Implemented CT_Font and tested the serialization.
git-svn-id: http://npoi.googlecode.com/svn/trunk@391 3d0cb27c-5942-9f1f-b0c8-b51bd5f61ba7

+++ ooxml/XSSF/UserModel/Extensions/XSSFCellFill.cs
@@ -69,7 +69,7 @@ namespace NPOI.XSSF.UserModel.Extensions
         {
             CT_PatternFill ptrn = EnsureCTPatternFill();
             CT_Color CT_Color = ptrn.IsSetBgColor() ? ptrn.bgColor : ptrn.AddNewBgColor();
-            CT_Color.indexed = index;
+            CT_Color.indexed = (uint)index;
         }
 
         /**
@@ -106,7 +106,7 @@ namespace NPOI.XSSF.UserModel.Extensions
         {
             CT_PatternFill ptrn = EnsureCTPatternFill();
             CT_Color CT_Color = ptrn.IsSetFgColor() ? ptrn.fgColor : ptrn.AddNewFgColor();
-            CT_Color.indexed = index;
+            CT_Color.indexed = (uint)index;
         }
 
         /**
